fix recursive macro expansion when encountering parens
testcase:

  #define INNER(X) foo(X)/bar(Y)
  #define OUTER(A, B) ()INNER(7)
  OUTER(1, 2)

get_macro_info is usually called recursively after encountering a
function-like macro, in which case it is correct to return when
a right parens was encountered which reduces bracelevel to 0.
however the first call is made with a "dummy" macro and the entire
macro body, so it is perfectly valid to encounter e.g. something
like (). this caused the function to prematurely exit before expanding
the macros after such a case.