implement difficult cornercase, expanded macro followed by parens
example:

  #define FOO(A) BAR
  #define BAR(X) bar, X
  FOO(X)(Y)

here FOO(X) expands to BAR. which is a function-type macro, and since
the next tokens here are (Y), BAR needs to be invoked with those.
since this cpp, unlike the others, doesn't fetch entire files into
memory and then turn them into tokenlists where insertions/additions
are relatively easy; instead it reads token by token from streams as
needed, some hacks were needed to make it work.
when we expand nested macros, in-memory streams are used which hold
just the contents of the parameters replaced and the macro body.
this makes it impossible to see if the original stream was followed
by (...). we basically need to store all tokenizers(with associated
streams) that were used so far until the current recursive expansion
level, and then go back in the list to check whether the next one
that's not at EOF can peek a '(' token, and if so, use that tokenizer
/stream to parse macro arguments.
unfortunately that's not all, we also need to remove the consumed
elements from the stream in case it's an in-memory one. the original
file-based stream is not seekable and is simply advanced to the right
position.
that's done by using the current offset of the stream as the end marker
for substitution, rather than the previously used token count.
this simply skips over the sections of the stream that have been
consumed by recursive calls.

having this implemented, we can finally preprocress preproc.c
with itself and the produced .i file compiles flawlessly with GCC.
i guess one could call that self-hosting (tested with musl libc).