tokenizer: do not join leading +/- with numeric literals
even though they could technically be part of the literal, it could
also have been used as an arithmetic operator. this causes an
ambiguity which can only be solved by having much more context av-
ailable. this broke evaluation of conditions.
for our purposes it's sufficient to simply emit the token as such
instead. the evaluator then figures out how it's supposed to be
interpreted.