properly catch and propagate buffer overflow errors
tokenizer_next is now defined as a macro checking for TT_OVERFLOW
tokentype. it does not fail on TT_UNKNOWN, unlike x_tokenizer_next.

there are only a few places where we actually need to care about
unknown tokens (e.g. in conditionals and function macro argument
parsing), but overflow needs to be treated as an error from anywhere.